Add an icon showcase to the home screen and change the supported hash functions

+++ lib/ui/screens/common/blocs/icon_event.dart
@@ -17,7 +17,7 @@ final class IconHashFunctionChanged extends IconEvent {
 final class IconGridTypeChanged extends IconEvent {
   IconGridTypeChanged({required this.gridType});
 
-  final double gridType;
+  final int gridType;
 }
 
 final class IconSizeChanged extends IconEvent {
